Move InputUtil to steps package where it is used
