+++ subprojects/execution/src/main/java/org/gradle/internal/execution/steps/InputUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.execution.impl;
+package org.gradle.internal.execution.steps;
 
 import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.api.UncheckedIOException;
